test(domain): 로또 객체 단위 테스트 코드 작성
테스트코드 목록:
  - 당첨 번호 개수 카운팅 기능 테스트
  - 보너스 번호 있는지 확인하는 기능 테스트
